add full docker support
config files like application.conf and formatting files are no longer included in the image, you should use volumes to link the directory on your local machine with these files to the /data/ folder on the docker container. this does mean that all extrenal config, format, etc. files are now looked for in /data/ not the current directory.

there is full integration between docker and gradle (so nice) thanks to bmuschko's gradle plugin.

use task createImage to prepair a dockerfile, create it, and create an image for it. pushImage unfortunatly isn't working yet due to authentication issues.