fix a bug where the wrong channel id was searched in the description
previously the description was searched for instances of ...tube.com/channel/{channel_id} however it was using the channel id of the channel were where polling and not of the channel we were supposed to be searching for.

a new entry was added to the bots.conf, "target". this holds a new field "channelId" (describing the channel id of the channel you are searching for), the already existing "youtubename" entry was moved to this entry as well.