add a test for youtube links not in our history
we were not testing if youtube links foudn in a description but not in our history were properly getting edited, this will test this mechenic. 

Because you need a youtube API key to make API requests this does not test the interaction with the Youtube Data API and just assumes it returns the correct data.