add default entry template
this template is located in the UserDefaults.conf file and can be changed (just like all the defaults) in the bots.conf. 

it add's a default template for every new bot instance, default settings include:
- intveral is set to 5
- listenertype is set to api
- pollertype is set to new video
- it uses the default comment settings with the only comment rule being the new line fix for reddit

it also adds the youtube api key when it is specified in the defaults object. 

all default values are taken from the defaults object, so if you want to change a default value in this entry template you only have to specify it once in the defaults.