add optional name config entry and enabled flag
the name entry is just metadata, it makes it a bit easier to find a specific instance in the logs.

the enabled flag does what it suggests and allows you to toggle on or off an instance, by default it is set to true but this can be changed in the defaults. 

when an instance is disabled no coresponding listener will be created and a warning message will be printed in the logs to alert the user.