fix a bug where the bot got overwhelmed
when one listener finds multiple new videos at once it processes the newest one first and the oldest one last causing the latest video date to be set to the oldest of the new videos instead of the newest.

we now reverse the list of new videos if it contains more than one entry causing the oldest entries to be handled first. We now also check when setting the latest video date if the "new" date is actually after the current latest video date